+++ doc/python/pages/changelog.rst
@@ -127,8 +127,9 @@ Changelog
 -   Exposed the :ref:`text` library
 -   Exposed the minimal interface of :ref:`utility.ConfigurationGroup` and
     :ref:`utility.Configuration`
--   Exposed :ref:`pluginmanager.AbstractManager.set_preferred_plugins()`, the
-    base :ref:`pluginmanager.AbstractPlugin` class and
+-   Exposed :ref:`pluginmanager.AbstractManager.set_preferred_plugins()`,
+    :ref:`pluginmanager.AbstractManager.register_external_manager()`, the base
+    :ref:`pluginmanager.AbstractPlugin` class and
     :ref:`pluginmanager.PluginMetadata`
 -   Fixed issues with an in-source build (see :gh:`mosra/magnum-bindings#13`)
 -   All CMake build options are now prefixed with ``MAGNUM_``. For backwards